[#137] 장소 북마크 api 및 UI 연결

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/dto/remote/response/bookmark/PlaceBookmarkData.kt
@@ -8,5 +8,5 @@ data class PlaceBookmarkData(
     val disability: List<String>,
     val image: String,
     val name: String,
-    val placeId: Int
+    val placeId: Long
 )
