+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/PlanBookmark.kt
@@ -3,7 +3,7 @@ package kr.tekit.lion.daongil.domain.model
 data class PlanBookmark (
     val image: String,
     val name: String,
-    val planId: Int,
+    val planId: Long,
     val profileImg: String,
     val title: String
 )
